feat: ଘ( ･ω･)_/ﾟ･:*:･｡☆ реализация метода встраивания ЦВЗ посредством дублирования точек feat 1: добавлена полная реализация (*) todo 1: комментирование методов кода todo 2: тестирование атак делециями/инсерциями

+++ DigitalWaterMarkApp/WaterMark/WaterMark.cs
@@ -73,16 +73,16 @@ namespace DigitalWaterMarkApp
             }
         }
 
-        private static int[] GetBinaryFromInteger(int length, int toBase, int number) {
+        private static int[] GetBinaryFromInteger(int length, int toBase, long number) {
             var source = ConvertToBase(number, toBase).PadLeft(length, '0');
             return source.Select(c => c - '0').ToArray();
         }
 
-        public static String ConvertToBase(int value, int toBase) => Convert.ToString(value, toBase);
+        public static String ConvertToBase(long value, int toBase) => Convert.ToString(value, toBase);
 
         public int ConvertToDecimal() => Convert.ToInt32(string.Join("", this.Items.Select(item => item.WMValue)), 2);
 
-        public static WaterMark ConvertToWaterMark(int value) {
+        public static WaterMark ConvertToWaterMark(long value) {
 
 
             int binaryLength = ConvertToBase(value, 2).Length;
